Make console logs viewable on browsers not supporting `console.log`.
### What
Add console logs to the console box and wrapped usage of `console.log` so it's not called if the browser doesn't support it.

### Why
Moving to non-jQuery I want to be able to easily test this on lots of old browsers, some that don't support console.log and on mobile devices that often don't let you view the JavaScript console.